requirements.txt
pandas
seaborn
pyarrow
matplotlib
fastparquet
qdrant_client
python-dotenv
simple-term-menu
sentence-transformers